+++ packages/devtools_app/lib/src/charts/flame_chart.dart
@@ -13,7 +13,7 @@ import 'package:flutter/services.dart';
 import '../analytics/analytics.dart' as ga;
 import '../primitives/auto_dispose_mixin.dart';
 import '../primitives/extent_delegate_list.dart';
-import '../primitives/linked_scroll_controller.dart';
+import '../primitives/flutter_widgets/linked_scroll_controller.dart';
 import '../primitives/trees.dart';
 import '../shared/common_widgets.dart';
 import '../shared/dialogs.dart';
