+++ packages/devtools_app/lib/src/debugger/codeview.dart
@@ -14,7 +14,7 @@ import 'package:vm_service/vm_service.dart' hide Stack;
 
 import '../config_specific/logger/logger.dart';
 import '../primitives/auto_dispose_mixin.dart';
-import '../primitives/linked_scroll_controller.dart';
+import '../primitives/flutter_widgets/linked_scroll_controller.dart';
 import '../shared/common_widgets.dart';
 import '../shared/dialogs.dart';
 import '../shared/globals.dart';
