+++ packages/devtools_app/lib/src/performance/flutter_frames_chart.dart
@@ -10,7 +10,7 @@ import 'package:provider/provider.dart';
 import '../analytics/analytics.dart' as ga;
 import '../analytics/constants.dart' as analytics_constants;
 import '../primitives/auto_dispose_mixin.dart';
-import '../primitives/linked_scroll_controller.dart';
+import '../primitives/flutter_widgets/linked_scroll_controller.dart';
 import '../shared/banner_messages.dart';
 import '../shared/common_widgets.dart';
 import '../shared/globals.dart';
