+++ packages/devtools_app/lib/src/performance/legacy/timeline_flame_chart.dart
@@ -16,8 +16,8 @@ import 'package:provider/provider.dart';
 import '../../analytics/constants.dart' as analytics_constants;
 import '../../charts/flame_chart.dart';
 import '../../primitives/auto_dispose_mixin.dart';
+import '../../primitives/flutter_widgets/linked_scroll_controller.dart';
 import '../../primitives/geometry.dart';
-import '../../primitives/linked_scroll_controller.dart';
 import '../../primitives/trace_event.dart';
 import '../../shared/common_widgets.dart';
 import '../../shared/notifications.dart';
