+++ packages/devtools_app/lib/src/primitives/flutter_widgets/README.md
@@ -1,0 +1,1 @@
+This folder contains code forked from https://github.com/google/flutter.widgets.
