+++ packages/devtools_app/lib/src/shared/table.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart' hide TableRow;
 import 'package:flutter/services.dart';
 
 import '../primitives/auto_dispose_mixin.dart';
-import '../primitives/linked_scroll_controller.dart';
+import '../primitives/flutter_widgets/linked_scroll_controller.dart';
 import '../primitives/trees.dart';
 import '../ui/colors.dart';
 import '../ui/search.dart';
